Refactor GeometricElement.Intersects into GeometricElement class.
